Renaming the file name
